@@ -1,34 +1,30 @@
+import { Routes, Route } from "react-router-dom";
+import Home from "./pages";
+import Blog from "./pages/blog";
+import Resume from "./pages/resume";
+import { ThemeProvider } from "styled-components";
 import {
-  BrowserRouter as Router,
-  Switch,
-  Route
-} from 'react-router-dom'
-import Home from './pages'
-import Blog from './pages/blog'
-import Resume from './pages/resume'
-import { ThemeProvider } from 'styled-components'
-import {
-  // Nature,
-  // Traditional,
-  Dark,
-} from './constants/colors'
+	// Nature,
+	// Traditional,
+	Dark,
+} from "./constants/colors";
 
 const theme = {
-  main: Dark,
-}
+	main: Dark,
+};
 
 function App() {
-  return (
-    <Router>
-      <Switch>
-        <ThemeProvider theme={theme}>
-          <Route path="/" component={Home} exact></Route>
-          <Route path="/resume" component={Resume} exact></Route>
-          <Route path="/blog" component={Blog} exact></Route>
-        </ThemeProvider>
-      </Switch>
-    </Router>
-  );
+	return (
+        // <div className="app">
+			<ThemeProvider theme={theme}>
+                <Routes>
+					<Route path="/" element={<Home/>}></Route>
+					<Route path="/resume" element={<Resume/>}></Route>
+					<Route path="/blog" element={<Blog/>}></Route>
+				</Routes>
+            </ThemeProvider>
+        // </div>
+	);
 }
 
 export default App;
