@@ -8,7 +8,7 @@ import {
 	FaReact,
 	FaAngular,
 	FaSass,
-	FaDatabase,
+    FaDatabase
 } from "react-icons/fa";
 
 import {
@@ -17,6 +17,8 @@ import {
 	SiLinux,
 	SiJava,
 	SiJavascript,
+    SiTypescript,
+    SiAuth0
 } from "react-icons/si";
 
 // import {
@@ -106,5 +108,14 @@ export const SKILL_ICON_COMP_MAP = {
 		icon: <FaSass />,
 		label: "Sass",
 		color: "#CD6799",
-	},
+    },
+    typescript: {
+        icon: <SiTypescript />,
+        label: "Typescript",
+        color: "#3178C6"
+    },
+    oauth: {
+        icon: <SiAuth0 />,
+        label: "Oauth2.0"
+    }
 };
