@@ -6,11 +6,14 @@ const data = {
     topLine: 'About me',
     headline: 'Hello, Gurpreet here!!',
     description: [
-    'I am a Software Engineer and I am passionate about coding. Starting my career back in 2013 with C++ and ' + 
-    'working in some core backend technologies for few years thereafter, I started getting inclined towards ui. ' +
-    'My last couple of years are mostly dedicated towards javascript and frameworks like react.',
-    'By the end of 2020, I decided to take a career break and dedicate some time to myself and ' +
-    'to do some coding just for fun ;)'
+    'An engineer at heart, passionate for programming and problem solving, started my career with C++ in the telecom domain, coding for protocols ' + 
+    'like Diameter following 3gpp specifications.',
+    'Later spent some years in the ERP domain, learning the basics of the finance and project ' +
+    'accounting modules in Oracle doing deep dives into the ocean of sql and pl/sql, doing handson on vast oracle tools like ODI, ADF, E-biz suite, Oracle Apex etc.', 
+    'Always curious and daring to take on new challenges, while at Oracle I volunteered for a project way out of my comfort zone which was front end heavy. ' +
+    'It was there when I first met Javascript, Html, Css and it was this feeling of instant gratification that I had the crush on the front end <3 for the first time in my life ;)',
+    'I grew more curious and wanted to know more about her, ahm..  sorry I mean the front end ;). Therefore, in the last 4 years I invested in job roles that require core ' + 
+    'softare engineering skills but are more inclined towards development of user interfaces.'
     ],
     buttonLabel: 'Contact',
     imgStart: false,
